chat_room/requirements.txt
asgiref==3.7.2
autopep8==2.3.2
bleach==6.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.8
cloudinary==1.44.0
colorama==0.4.6
cryptography==42.0.5
defusedxml==0.7.1
dj-database-url==2.1.0
Django==5.0.3
django-allauth==0.61.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-debug-toolbar==5.2.0
django-seed==0.3.1
djangorestframework==3.15.1
Faker==24.4.0
gprof2dot==2025.4.14
gunicorn==22.0.0
h11==0.14.0
idna==3.7
iniconfig==2.1.0
oauthlib==3.2.2
packaging==24.0
pillow==10.2.0
pluggy==1.6.0
psycopg2==2.9.9
pycodestyle==2.13.0
pycparser==2.22
PyJWT==2.8.0
pytest==8.3.5
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytube==15.0.0
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==0.10.1
six==1.16.0
sqlparse==0.4.4
toposort==1.10
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.20
uvicorn==0.34.0
webencodings==0.5.1
whitenoise==6.8.2
